Address #496

+++ includes/trait-wp-document-revisions-file-handler.php
@@ -558,7 +558,7 @@ trait WP_Document_Revisions_File_Handler {
 	public function filename_rewrite( array $file ): array {
 		// verify this is a document.
 		// phpcs:ignore WordPress.Security.NonceVerification.Missing
-		if ( ! isset( $_POST['post_id'] ) || ! $this->verify_post_type( sanitize_text_field( wp_unslash( $_POST['post_id'] ) ) ) ) {
+		if ( ! isset( $_POST['post_id'] ) || ! $this->verify_post_type( (int) sanitize_text_field( wp_unslash( $_POST['post_id'] ) ) ) ) {
 			self::$doc_image = true;
 			return $file;
 		}
